+++ components/o3-form/src/tsx/fieldComponents/Feedback.tsx
@@ -1,0 +1,5 @@
+import type {FeedBackProps} from '../../types';
+
+export const Feedback = ({message, type}: FeedBackProps) => {
+	return <div className={`feedback ${type}`}>{message}</div>;
+};
